Added: 파티셔닝 Job을 위한 구성 작업

+++ src/main/kotlin/com/example/batch/domain/User.kt
@@ -1,3 +1,5 @@
 package com.example.batch.domain
 
-data class User(val id: Int, val name: String)
+data class User(
+    val id: Long,
+    val name: String)
